+++ src/node/find.rs
@@ -64,10 +64,10 @@ impl<S> Node<S> {
     }
 
     fn find_end_wildcard(&self, template: &mut Template, name: &str) -> Option<&NodeData> {
-        if let Some(child) = &self.end_wildcard {
-            if child.state.name == name {
-                return child.find(template);
-            }
+        if let Some(child) = &self.end_wildcard
+            && child.state.name == name
+        {
+            return child.find(template);
         }
 
         None